refactor: lint cleanup wave 7 — line-preserving auto-fixes (77 errors) (#425)
Apply line-preserving ruff auto-fixes (no line-number shifts → ledger tests stay green):

- **RUF059**: unused tuple unpack → `_var` (46)
- **RUF046**: unnecessary `int(...)` cast (18)
- **RUF021**: parenthesize `(a and b) or (c and d)` for explicit grouping (7 leftover)
- **SIM118**: drop `.keys()` from iteration (5 leftover)
- **SIM910**: remove `dict.get(k, None)` (1 leftover)

70 insertions / 70 deletions across 44 files. Ledger sweep green
(2840 passed / 9 skipped / 11190 deselected).

Skipped W191 (598 in 3 files: `terminal_feed_state.py`, `terminal_resolution_state.py`,
`streamlit_databento_volatility_screener.py`) — `ruff check` does not auto-fix W191
and `ruff format` would reflow content (line-shift risk). Defer to a separate PR.

+++ open_prep/psr_robust.py
@@ -104,8 +104,8 @@ def _winsorize(values: Sequence[float], alpha: float) -> list[float]:
         raise ValueError(f"alpha must be in (0, 0.5), got {alpha}")
     sorted_vals = sorted(values)
     n = len(sorted_vals)
-    lo_idx = max(0, int(math.floor(alpha * n)))
-    hi_idx = min(n - 1, int(math.ceil((1.0 - alpha) * n)) - 1)
+    lo_idx = max(0, math.floor(alpha * n))
+    hi_idx = min(n - 1, math.ceil((1.0 - alpha) * n) - 1)
     lo = sorted_vals[lo_idx]
     hi = sorted_vals[hi_idx]
     return [min(max(v, lo), hi) for v in values]
